requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
docutils==0.14
ecdsa==0.13
future==0.17.1
html5lib==1.0.1
idna==2.7
Markdown==3.0.1
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.0
python-jose==3.0.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
webencodings==0.5.1
